+++ .gitignore
@@ -1,0 +1,2 @@
+venv
+__pycache__
