+++ Pipfile
@@ -49,6 +49,7 @@ flask-cors = "*"
 flask-script = "*"
 flask-migrate = "*"
 faker = "*"
+psycopg2 = "*"
 
 [requires]
 python_version = "3.8"
