+++ app/__init__.py
@@ -2,9 +2,15 @@ from flask import Flask, jsonify
 from flask_sqlalchemy import SQLAlchemy
 from flask_cors import CORS, cross_origin
 import os
+from config import app_config
 #create app method
+#run env
 app = Flask(__name__)
-db = SQLAlchemy()
+app.config['CORS_HEADERS'] = 'Content-Type'
+app.config['SQLALCHEMY_DATABASE_URI'] = os.environ.get('SQLALCHEMY_DATABASE_URI')
+app.config.from_object(app_config[os.environ.get('FLASK_ENV', 'production')])
+app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
+db = SQLAlchemy(app)
 # db.init_app(app)
 cors = CORS(app, support_credentials=True, resources={
     r"/*": {
@@ -12,33 +18,8 @@ cors = CORS(app, support_credentials=True, resources={
     }
 })
 
-# app.config['CORS_HEADERS'] = 'Content-Type'
-# app.config['SQLALCHEMY_DATABASE_URI'] = os.environ.get('DATABASE_URL')
-# app.config.from_object(config['FLASK_ENV'])
-# app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
 
 
-class Development(object):
-    DEBUG = True
-    TESTING = False
-    SQLALCHEMY_DATABASE_URI = os.getenv('SQLALCHEMY_DATABASE_URI')
-
-class Production(object):
-    DEBUG = True
-    TESTING = False
-    SQLALCHEMY_DATABASE_URI = os.getenv('SQLALCHEMY_DATABASE_URI')
-    SQLALCHEMY_TRACK_MODIFICATIONS = False
-
-class Testing(object):
-    TESTING = True
-    SQLALCHEMY_DATABASE_URI = os.getenv('SQLALCHEMY_TEST_DATABASE_URI')
-    SQLALCHEMY_TRACK_MODIFICATIONS = False
-
-app_config = {
-    'development': Development,
-    'production': Production,
-    'testing': Testing
-}
 
 
 from app import routes
