+++ readme.md
@@ -1,0 +1,9 @@
+Notes to make a virtual environment
+```
+python -m venv venv
+pip install -r requirements.txt
+```
+```
+source venv/bin/activate
+```
+deactivate
